v2.5.0: Code cleanup before reader and parser final optimizations

+++ includes/utils.h
@@ -4,6 +4,6 @@
 # include "bsq.h"
 
 t_bool	is_print(char c);
-int		ft_fast_atoi_n(char *str, size_t n);
+size_t	ft_fast_atoi_n(char *str, size_t n);
 
 #endif
